requirements.txt
asgiref==3.5.0
Django==3.1.5
pytz==2021.3
sqlparse==0.4.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.7.0
psycopg2-binary==2.8.6
django-filter==21.1